+++ tests/testthat/test_harmonization.R
@@ -25,6 +25,7 @@ test_that("trait harmonization can be done", {
                   DOI ="xxx",
                   Priority = 3)
   db_post <- traits4models::harmonize_taxonomy_WFO(db_var, WFO_file, progress = FALSE)
+  expect_s3_class(db_post, "data.frame")
 })
 
 test_that("harmonization checks are ok",{
@@ -32,4 +33,10 @@ test_that("harmonization checks are ok",{
   testthat::skip_on_ci()
   expect_type(check_harmonized_trait_dir(harmonized_trait_path, verbose = FALSE), "logical")
   expect_type(check_harmonized_allometry_dir(harmonized_allometry_path, verbose = FALSE), "logical")
+})
+
+test_that("harmonized data can be loaded",{
+  testthat::skip_on_cran()
+  testthat::skip_on_ci()
+  expect_s3_class(get_trait_data(harmonized_trait_path, "Al2As"), "data.frame")
 })