+++ NEWS.md
@@ -1,9 +1,11 @@
 # traits4models 0.2.1
 * New columns 'OriginalReference' and 'OriginalDOI' now possible in harmonized trait data
 * Functions get_trait_data() and get_taxon_data() now return 'OriginalReference' and 'OriginalDOI' if available 
-* New trait definitions 'StemEPS', 'CCleaf', 'CCsapwood' and 'CCfineroot' and leaf phenology traits
+* New trait definitions
 * New functions 'check_harmonized_trait_dir()' and 'check_harmonized_allometry_dir()'
 * Trait data can be harmonized in wide and long format
+* New fields to check harmonization: 'AcceptedValues'
+* Added checkVersion to store package version used for harmonization
 
 # traits4models 0.2.0
 * New columns 'MininumValue' and 'MaximumValue' added to HarmonizedTraitDefinition